bubble_my_tea/requirements.txt
asgiref==3.8.1
astroid==3.1.0
bcrypt==4.1.2
dill==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
isort==5.13.2
mccabe==0.7.0
mysqlclient==2.2.4
pillow==10.3.0
platformdirs==4.2.0
PyJWT==2.8.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
sqlparse==0.5.0
tomlkit==0.12.4
